requirements.txt
pyyaml
requests
beautifulsoup4
pytz
